requirements.txt
flask
gunicorn
google-generativeai
openai
python-dotenv
requests
sqlalchemy
